2020-03-09 Update: Added Rubocop integration

+++ lib/minfo/memory_shot.rb
@@ -4,29 +4,31 @@ module Minfo
   class OSInfoUnavailable < StandardError; end
 
   class MemoryShot
-    @@memory_shot = Hash.new
+    def initialize
+      @memory_shot = Hash.new
+    end
 
     def total
       prepare_memory_shot
-      @@memory_shot['total']
+      @memory_shot['total']
     end
 
     def prepare_memory_shot
-      raise OSInfoUnavailable unless File.exist?('/proc/meminfo')
+      if @memory_shot.empty?
+        raise OSInfoUnavailable, 'Cannot find or read /proc/meminfo' unless File.exist?('/proc/meminfo')
 
-      if @@memory_shot.empty?
         total = 0
         free = 0
         File.open('/proc/meminfo').read.each_line { |line|
           if line.start_with?('MemTotal:')
             total = extract_memory_value(line)
-            @@memory_shot['total'] = MemoryUnit.new(total)
+            @memory_shot['total'] = MemoryUnit.new(total)
           elsif line.start_with?('MemFree:')
             free = extract_memory_value(line)
-            @@memory_shot['free'] = MemoryUnit.new(free)
+            @memory_shot['free'] = MemoryUnit.new(free)
           end
         }
-        @@memory_shot['used'] = MemoryUnit.new(total - free)
+        @memory_shot['used'] = MemoryUnit.new(total - free)
       end
     end
 
@@ -36,12 +38,12 @@ module Minfo
 
     def free
       prepare_memory_shot
-      @@memory_shot['free']
+      @memory_shot['free']
     end
 
     def used
       prepare_memory_shot
-      @@memory_shot['used']
+      @memory_shot['used']
     end
   end
 end
